if a new object appears in the searchradius of an existing one, the ids
get swapped and therefore the two change position which is especially visible when smoothing

this commit changes the matching delegate to output a float rather than a bool.
where the bool was used to determine if the 2 object match, the float describes a similarity.
the following rules get implicitly applied: the smaller the value, the more similar.
if the value is above 1, it will never taken into account - like the bool before.

NOTE:
one more optimization could be made:
in case a new object appears exactly between two ghosts, the 'younger' ghost should be chosen.